feat: Optimization of asynchronous requests and their processing

+++ Api/delete-link-director-movie.php
@@ -1,12 +1,36 @@
 <?php
-// Objectif: supprimer le lien entre un film est un réalisateur
-// Récupérez les ID de director et movie à partir des données de la requête AJAX
-$directorId = $_POST['director_id'];
-$movieId = $_POST['movie_id'];
+require_once $_SERVER['DOCUMENT_ROOT'] . '/config.php';
 
-$pdo = new PDO('mysql:host=db;dbname=moviz_db', 'test', 'test');
-$stmt = $pdo->prepare('DELETE FROM movie_director WHERE director_id = :director_id AND movie_id = :movie_id');
-$stmt->execute(['director_id' => $directorId, 'movie_id' => $movieId]);
+$data = json_decode(file_get_contents('php://input'), true);
 
-// Envoyez une réponse au script JavaScript
-http_response_code(200);
+if (!isset($data['director_id']) || !isset($data['movie_id'])) {
+    http_response_code(400);
+    echo json_encode(['error' => 'Missing required fields']);
+    exit;
+}
+
+$directorId = (int)$data['director_id'];
+$movieId = (int)$data['movie_id'];
+
+try {
+
+    $pdo = new PDO("mysql:dbname=" . _DB_NAME_ . ";host=db;charset=utf8mb4", _DB_USER_, _DB_PASSWORD_);
+    $pdo->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
+
+    $stmt = $pdo->prepare('DELETE FROM movie_director WHERE director_id = :director_id AND movie_id = :movie_id');
+
+    $stmt->bindParam(':director_id', $directorId, $pdo::PARAM_INT);
+    $stmt->bindParam(':movie_id', $movieId, $pdo::PARAM_INT);
+    $result = $stmt->execute();
+
+    if (!$result) {
+        http_response_code(500);
+        echo json_encode(['error' => 'Failed to insert data into database']);
+        exit;
+    } else {
+        echo json_encode(['success' => 'Suppression du réalisateur effectuée ']);
+        http_response_code(200);
+    }
+} catch (PDOException $e) {
+    error_log($e->getMessage());
+}
